Create Formatted File Name

Create a standardized file name of the form: TXX-Description-MM_DD_YYYY.ext where 'T' is a leading tag, 'XX' is a file number, 'Description' is a human-readable label, and 'ext' is a file extension.

make_file_name(
  description,
  extension,
  tag = NULL,
  number = NULL,
  file_date = NULL,
  additional = NULL,
  date_format = "%m_%d_%Y",
  exclude = "",
  remove = FALSE
)

Arguments

description
A human-readable label, with words preferably separated by underscores.

extension
A file extension, such as 'RData', 'R', 'txt', 'pdf'.

tag
A leading tag; If no value is provided, automatically set based on the file extension. Automatic assignments are...

'S' for extension R;

'D' for extensions RData and csv;

'T' for extension txt;

'W' for extension docx;

'P' for extension pptx;

'F' for extensions pdf, jpg, jpeg, and png.

number
A file number. If no value is provided, automatically set based on number of files in current folder with matching tags.

file_date
The date to include. If no value is provided, the current date is used.

additional
Additional text to include following the date. If provided, is preceded by a '-'.

date_format
The format to use for the current date, defaults to 'MM_DD_YYYY'.

exclude
A vector of file names to exclude when

remove
Logical; if TRUE, attempts to locate previous versions of the outputted file name (i.e., same name but earlier dates) and remove them from the current folder.

Value

A character string.

Examples

# Different file types
make_file_name("Example", "RData")

#> [1] "D01-Example-05_20_2021.RData"

make_file_name("Example", "pdf")

#> [1] "F01-Example-05_20_2021.pdf"

make_file_name("Example", "docx")

#> [1] "W01-Example-05_20_2021.docx"

# User-specified tags and numbers
make_file_name("Example", "RData", tag = "R", number = "02")

#> [1] "R02-Example-05_20_2021.RData"

# Additional text
make_file_name("Example", "RData", additional = "v.1.0.0")

#> [1] "D01-Example-05_20_2021-v.1.0.0.RData"